requirements.txt
paho-mqtt
numpy
keras
pandas
python-dateutil
datetime
requests
scikit-learn